pyproject.toml
[build-system]
requires = ["setuptools>=60", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "./src/steering_metrics/_version.py"
local_scheme = "no-local-version"

[tool.setuptools]
package-dir = {"" = "src"}
zip-safe = false
include-package-data = true

[project]
name = "mvp_mpc"
version = "0.0.1"
description = "T"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "attrs==25.4.0",
    "black==25.1.0",
    "boto3==1.40.59",
    "botocore==1.40.59",
    "cachetools==5.5.2",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "db-dtypes>=1.4.3",
    "fonttools==4.60.1",
    "frozenlist==1.8.0",
    "google-cloud-bigquery>=3.38.0",
    "idna==3.11",
    "ipykernel>=6.29.5",
    "jmespath==1.0.1",
    "kiwisolver==1.4.9",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "matplotlib==3.10.7",
    "mdurl==0.1.2",
    "multidict==6.7.0",
    "myPyllant==0.9.6",
    "numpy==2.3.4",
    "packaging==25.0",
    "pandas==2.3.3",
    "pandas-gbq>=0.29.2",
    "pillow==12.0.0",
    "plotly>=6.3.1",
    "pre-commit>=4.2.0",
    "propcache==0.4.1",
    "pydantic==2.12.3",
    "pydantic_core==2.41.4",
    "pyarrow>=21.0.0",
    "Pygments==2.19.2",
    "pymelcloud>=2.11.0",
    "pyparsing==3.2.5",
    "pytest>=7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "requests==2.32.5",
    "rich==14.2.0",
    "ruff==0.11.3",
    "s3transfer==0.14.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.15.3",
    "seaborn==0.13.2",
    "selenium>=4.30.0",
    "setuptools>=78.1.0",
    "six==1.17.0",
    "thermiaonlineapi>=6.3.1",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "yarl==1.22.0",
    "arrow>=1.4.0",
]

[tool.ruff]
line-length = 100

# lint.select = [
#     "C4", "COM", "D", "E", "EM", "ERA", "F", "I",
#     "NPY", "PD", "PIE", "PL", "PT", "RET", "RSE",
#     "RUF", "SIM", "TID", "UP", "W"
# ]

# Configure quote style to prefer double quotes
lint.pydocstyle.convention = "google"
lint.flake8-quotes.inline-quotes = "double"
lint.flake8-quotes.multiline-quotes = "double"
lint.flake8-quotes.docstring-quotes = "double"

lint.ignore = [
    "D205",  # 1 blank line required between summary line and description (autofix not working)
    "D104",  # Missing docstring in public package
    "COM812",  # Trailing comma in multiline collection
]

# Enable automatic fixing
fix = true

[tool.bandit]
skips = ["B108", "B307", "B404", "B602"]

[tool.bandit.assert_used]
skips = ["**/test_*.py", "**/*_test.py"] ## TBD
